Fixes for stringUtils issues * tS macro from stringUtils is used in various user projects compiled in C++98 mode meaning that new stringUtils fail to compile - therefore moved more code-gen specific stuff out into 'codeGenUtils' * Fixed more warnings * Removed more tS usage from GeNN library
